Filewatch on linux (#8)  - Updated transcode tools version for linux as it now includes libjnotify.so (x64) and libjnotify-x86.so which is required for jnotify to work on linux  - Copy libjnotify.so (and libjnotify-x86.so) to standard library path in core to make jnotify work from eclipse  - Include libjnotify.so and libjnotify-x86.so when packaging
It has only been tested on ubuntu 13.10 x64. No delete notifications were raised when a file in a watched folder was being deleted (this works on win and osx). If running on a x86 OS, libjnotify-x86.so will have to be renamed to libjnotify.so in order to work, but this hasn't been tested.